Add seek backwards
This function adds the option to seek backwards an indicated time_delta.

This is used to emulate a 'back' button that works well for music (short songs) as well as audiobooks (long chapters).

This is achieved by the folowing logic:

a) song already running longer than time_delta: the song position is
  reverted back a time delta -> enables to seek within long audiobooks

b) song running shorter than time_delta: Go to the previous song. It his is longer than time_delta, go to the position one time-delta from the end.
   Otherwise go to the beginning.
   -> enables to seek back to the previous song